update currency code method
