+++ tests/http-server/units/prices.spec.js
@@ -286,67 +286,6 @@ describe('Units prices', function () {
       const res = await response.json();
       expect(res).to.have.property('code', '#missingPrice');
     });
-    it('POST /hotels/:hotelAddress/units/:unitAddress/currencyCode. Expect 200', async () => {
-      const body = JSON.stringify({
-        password: config.get('password'),
-        code: 948,
-      });
-      let response = await fetch(`http://localhost:3000/hotels/${config.get('testAddress')}/units/${config.get('unitAddress')}/currencyCode`, {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json',
-        },
-        body,
-      });
-      expect(response).to.be.ok;
-      expect(response).to.have.property('status', 200);
-      response = await fetch(`http://localhost:3000/hotels/${config.get('testAddress')}`, {
-        method: 'GET',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json',
-          'Content-Length': Buffer.byteLength(body),
-        },
-        body,
-      });
-      const { hotel } = await response.json();
-      expect(hotel.units[config.get('unitAddress')]).to.have.property('currencyCode', 'CHW');
-    });
-    it('POST /hotels/:hotelAddress/units/:unitAddress/currencyCode. Expect 400 #missingPassword', async () => {
-      const body = JSON.stringify({
-        code: 948,
-      });
-      let response = await fetch(`http://localhost:3000/hotels/${config.get('testAddress')}/units/${config.get('unitAddress')}/currencyCode`, {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json',
-        },
-        body,
-      });
-      expect(response).to.be.ok;
-      expect(response).to.have.property('status', 400);
-      const resp = await response.json();
-      expect(resp).to.have.property('code', '#missingPassword');
-    });
-    it('POST /hotels/:hotelAddress/units/:unitAddress/currencyCode. Expect 400 #missingCode', async () => {
-      const body = JSON.stringify({
-        password: config.get('password'),
-      });
-      let response = await fetch(`http://localhost:3000/hotels/${config.get('testAddress')}/units/${config.get('unitAddress')}/currencyCode`, {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json',
-        },
-        body,
-      });
-      expect(response).to.be.ok;
-      expect(response).to.have.property('status', 400);
-      const resp = await response.json();
-      expect(resp).to.have.property('code', '#missingCode');
-    });
     it('POST /hotels/:hotelAddress/units/:unitAddress/specialPrice. Expect 200', async () => {
       const specialPrice = 70;
       let body = JSON.stringify({
